requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
pinecone-client==2.2.4
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.3
requests==2.31.0